This folder contains the CombinedScript
Combined the two scripts to work with a circuit. Also updated the script to output a full read of all sensors based on each condition. I believe this will be important when outputting the files through JSON as we can have a data point for every important change. The next version of this script will include the ArduinoJSON library for formatting, and any changes to logic that are necessary.
This portion has not been tested or refined yet as the circuit has not been fully built.