Working 1 Min messaging script
This script will be the version used to send data to our web application. It features conversions on the moisture sensor to show as relative percent. It also populates alerts if the moisture level is below 20%, or if the temperature is above 40C. There is approximately 10s delay from real time. The moisture has some accuracy issues, should be within ~10% accuracy towards the upper limit.